requirements.txt
azure-identity
# Flask[async]==2.3.2
openai
azure-search-documents
azure-storage-blob
python-dotenv
azure-cosmos
quart
uvicorn
aiohttp
gunicorn
pymongo
azure-keyvault-secrets